New features: inline broker request-response, temporary queues, improved docs and examples and better testing (#39)
* Examples Update + Code Refactor (#25)

* Removed StorageManager

* Added examples for OpenAI, Bedrock, Anthropic, and VertexAI

* Updating old examples (1/2)

* Updating old examples (2/2)

* Added support for temporary queue + UUID queue name (#26)

* Add assembly component and auto-generated documents (#27)

* Added the assembly component

* Auto-generated documents

* Added type check

* Update the cache service expiry logic + Update the assembly component to use cache expiry for timeout

* Moved assembly to the correct place

* Added MoA Example + UUID Invoke Function (#28)

* MoA example: Broadcast to multiple agents

* Added MoA event manager, added uuid invoke_function + test, updated auto-generated docs

* Added assembly layer to MoA example

* Update documentation for new users + Refactored component_input & source_expression (#29)

* Refactored component_input to input_selection

* Updated, added, and enhanced the documentation with new users in mind

* Refactored source_expression function to evaluate_expression (backward compatible)

* Added tips and tricks section + info and examples on custom modules

* tiny format update

* tiny update

* Fixed solace disconnection issues on shutting down (#30)

* Add RAG example for AI connector + delete action for vector index (#31)

* Added a RAG example for AI connector

* Added delete option to vectordb

* Changed id to ids

* chore: Refactor make_history_start_with_user_message method (#32)

Fix the method to not trim the first entry if it is a "system" role

* Keep history depth needs to be a positive integer and test refactor (#33)

* chore: Refactor clear_history_but_keep_depth method to handle negative depth values

* chore: small change to how this is solved

* chore: one more try

* refactor: move utils_for_test_files.py to solace_ai_connector module

* refactor: removed the orginal utils_for_test_files.py

* refactor: update import statements in test files

* refactor: add sys.path.append("src") to test files

* refactor: standardize import order and sys.path.append in test files

* refactor: a bit more test infrastructure changes

* feat: allow component_module to accept module objects directly

* feat: add types module import to utils.py

* test: add static import and object config test

* refactor: update test_static_import_and_object_config to use create_test_flows

* refactor: Improve test structure and remove duplicate test case

* fix: remove duplicate import of yaml module

* refactor: Modify test config to use dict instead of YAML string

* refactor: convert config_yaml from string to dictionary

* refactor: update static import test to use pass_through component

* test: Add delay component message passing test

* feat: add test for delay component message processing

* feat: Added a new test function (test_one_component) to make it very easy to just run some quick tests on a single input -> expected output tests on a single component

* feat: added input_transforms to the test_one_component so that input transforms can be tested with it

* chore: a bit of cleanup and new tests for test_one_component

* chore: rename test_one_component because it was being picked up as a test by the pytest scanner

* fix: fixed a typo

* Fix for anthropic example (#35)

* Updating version dependency (#37)

* Fixed url and file name in getting started (#38)

* Add guide for RAG (#39)

* Added guide for RAG

* update wording

* Added link to other docs from RAG guide (#40)

* chore: added a timeout setting for running component tests so that you can test situations where you don't expect any output (#34)

* AI-124: Add a feature to provide simple blocking broker request/response ability for components (#42)

* feat: add request_response_controller.py

* feat: implement RequestResponseFlowManager and RequestResponseController classes

* style: format code with black and improve readability

* feat: implement RequestResponseController for flow-based request-response handling

* feat: implement RequestResponseController for handling request-response patterns

* fix: import SolaceAiConnector for type checking

* refactor: restructure Flow class and improve code organization

* feat: implement multiple named RequestResponseControllers per component

* refactor: initialize request-response controllers in ComponentBase

* test: add request_response_controller functionality tests

* feat: finished implementation and added some tests

* refactor: rename RequestResponseController to RequestResponseFlowController

* refactor: rename RequestResponseController to RequestResponseFlowController

* refactor: some name changes

* fix: update test function names for RequestResponseFlowController

* refactor: more name changes

* Ed/req_resp_examples_and_fixes (#41)

* feat: Added a request_response_flow example and fixed a few issues along the way

* feat: Reworked the broker_request_response built-in ability of components to be simpler. Instead of having to have a defined flow and then name that flow, it will automatically create a flow with a single broker_request_response component in it. Now there is a straightforward interating function call to allow components to issue a request and get streaming or non-streaming responses from that flow.

* chore: fix the request_response example and remove the old one

* docs: add broker request-response configuration

* docs: added advanced_component_features.md

* docs: add broker request-response configuration details

* docs: add payload encoding and format to broker config

* docs: add cache service and timer manager to advanced_component_features.md

* docs: add configuration requirement for broker request-response

* docs: update broker request-response section with configuration info

* docs: a bit more detail about do_broker_request_response

* docs: add link to advanced features page in table of contents

* docs: add link to advanced features page

* docs: reorder table of contents in index.md

* docs: add custom components documentation

* docs: Remove advanced component features from table of contents

* docs: clean up a double inclusion of the same section

* docs: small example change

* chore: remove dead code

* chore: add some extra comments to explain some test code

* docs: Update description of STDIN input component

Update the description of the STDIN input component to clarify that it waits for its output message to be acknowledged before prompting for the next input. This change is made in the `stdin_input.py` file.

* chore: add is_broker_request_response_enabled method

* chore: Some changes after review

* feat: AI-129: add ability to specify a default value for a an environment variable in a .yaml config file (#43)

* DATAGO-85484 Bump min python version

---------

+++ src/solace_ai_connector/components/inputs_outputs/broker_request_response.py
@@ -3,6 +3,8 @@
 import threading
 import uuid
 import json
+import queue
+from copy import deepcopy
 
 # from typing import Dict, Any
 
@@ -63,6 +65,19 @@ info = {
             "description": "Expiry time for cached requests in milliseconds",
             "default": 60000,
         },
+        {
+            "name": "streaming",
+            "required": False,
+            "description": "The response will arrive in multiple pieces. If True, "
+            "the streaming_complete_expression must be set and will be used to "
+            "determine when the last piece has arrived.",
+        },
+        {
+            "name": "streaming_complete_expression",
+            "required": False,
+            "description": "The source expression to determine when the last piece of a "
+            "streaming response has arrived.",
+        },
     ],
     "input_schema": {
         "type": "object",
@@ -78,6 +93,16 @@ info = {
             "user_properties": {
                 "type": "object",
                 "description": "User properties to send with the request message",
+            },
+            "stream": {
+                "type": "boolean",
+                "description": "Whether this will have a streaming response",
+                "default": False,
+            },
+            "streaming_complete_expression": {
+                "type": "string",
+                "description": "Expression to determine when the last piece of a "
+                "streaming response has arrived. Required if stream is True.",
             },
         },
         "required": ["payload", "topic"],
@@ -115,6 +140,11 @@ class BrokerRequestResponse(BrokerBase):
         self.reply_queue_name = f"reply-queue-{uuid.uuid4()}"
         self.reply_topic = f"reply/{uuid.uuid4()}"
         self.response_thread = None
+        self.streaming = self.get_config("streaming")
+        self.streaming_complete_expression = self.get_config(
+            "streaming_complete_expression"
+        )
+        self.broker_type = self.broker_properties.get("broker_type", "solace")
         self.broker_properties["temporary_queue"] = True
         self.broker_properties["queue_name"] = self.reply_queue_name
         self.broker_properties["subscriptions"] = [
@@ -123,7 +153,13 @@ class BrokerRequestResponse(BrokerBase):
                 "qos": 1,
             }
         ]
-        self.connect()
+        self.test_mode = False
+
+        if self.broker_type == "solace":
+            self.connect()
+        elif self.broker_type == "test" or self.broker_type == "test_streaming":
+            self.test_mode = True
+            self.setup_test_pass_through()
         self.start()
 
     def start(self):
@@ -135,8 +171,16 @@ class BrokerRequestResponse(BrokerBase):
             self.reply_queue_name, [self.reply_topic], temporary=True
         )
 
+    def setup_test_pass_through(self):
+        self.pass_through_queue = queue.Queue()
+
     def start_response_thread(self):
-        self.response_thread = threading.Thread(target=self.handle_responses)
+        if self.test_mode:
+            self.response_thread = threading.Thread(
+                target=self.handle_test_pass_through
+            )
+        else:
+            self.response_thread = threading.Thread(target=self.handle_responses)
         self.response_thread.start()
 
     def handle_responses(self):
@@ -148,59 +192,74 @@ class BrokerRequestResponse(BrokerBase):
             except Exception as e:
                 log.error("Error handling response: %s", e)
 
+    def handle_test_pass_through(self):
+        while not self.stop_signal.is_set():
+            try:
+                message = self.pass_through_queue.get(timeout=1)
+                decoded_payload = self.decode_payload(message.get_payload())
+                message.set_payload(decoded_payload)
+                self.process_response(message)
+            except queue.Empty:
+                continue
+            except Exception as e:
+                log.error("Error handling test passthrough: %s", e)
+
     def process_response(self, broker_message):
-        payload = broker_message.get_payload_as_string()
-        if payload is None:
-            payload = broker_message.get_payload_as_bytes()
-        payload = self.decode_payload(payload)
-        topic = broker_message.get_destination_name()
-        user_properties = broker_message.get_properties()
+        if self.test_mode:
+            payload = broker_message.get_payload()
+            topic = broker_message.get_topic()
+            user_properties = broker_message.get_user_properties()
+        else:
+            payload = broker_message.get_payload_as_string()
+            if payload is None:
+                payload = broker_message.get_payload_as_bytes()
+            payload = self.decode_payload(payload)
+            topic = broker_message.get_destination_name()
+            user_properties = broker_message.get_properties()
 
         metadata_json = user_properties.get(
             "__solace_ai_connector_broker_request_reply_metadata__"
         )
         if not metadata_json:
-            log.warning("Received response without metadata: %s", payload)
+            log.error("Received response without metadata: %s", payload)
             return
 
         try:
             metadata_stack = json.loads(metadata_json)
         except json.JSONDecodeError:
-            log.warning(
-                "Received response with invalid metadata JSON: %s", metadata_json
-            )
+            log.error("Received response with invalid metadata JSON: %s", metadata_json)
             return
 
         if not metadata_stack:
-            log.warning("Received response with empty metadata stack: %s", payload)
+            log.error("Received response with empty metadata stack: %s", payload)
             return
 
         try:
             current_metadata = metadata_stack.pop()
         except IndexError:
-            log.warning(
+            log.error(
                 "Received response with invalid metadata stack: %s", metadata_stack
             )
             return
         request_id = current_metadata.get("request_id")
         if not request_id:
-            log.warning("Received response without request_id in metadata: %s", payload)
+            log.error("Received response without request_id in metadata: %s", payload)
             return
 
         cached_request = self.cache_service.get_data(request_id)
         if not cached_request:
-            log.warning("Received response for unknown request_id: %s", request_id)
+            log.error("Received response for unknown request_id: %s", request_id)
             return
+
+        stream = cached_request.get("stream", False)
+        streaming_complete_expression = cached_request.get(
+            "streaming_complete_expression"
+        )
 
         response = {
             "payload": payload,
             "topic": topic,
             "user_properties": user_properties,
-        }
-
-        result = {
-            "request": cached_request,
-            "response": response,
         }
 
         # Update the metadata in the response
@@ -221,14 +280,35 @@ class BrokerRequestResponse(BrokerBase):
                 "__solace_ai_connector_broker_request_reply_topic__", None
             )
 
-        self.process_post_invoke(result, Message(payload=result))
-        self.cache_service.remove_data(request_id)
+        message = Message(
+            payload=payload,
+            user_properties=user_properties,
+            topic=topic,
+        )
+        self.process_post_invoke(response, message)
+
+        # Only remove the cache entry if this isn't a streaming response or
+        # if it is the last piece of a streaming response
+        last_piece = True
+        if stream and streaming_complete_expression:
+            is_last = message.get_data(streaming_complete_expression)
+            if not is_last:
+                last_piece = False
+
+        if last_piece:
+            self.cache_service.remove_data(request_id)
 
     def invoke(self, message, data):
         request_id = str(uuid.uuid4())
 
         if "user_properties" not in data:
             data["user_properties"] = {}
+
+        stream = False
+        if "stream" in data:
+            stream = data["stream"]
+        if "streaming_complete_expression" in data:
+            streaming_complete_expression = data["streaming_complete_expression"]
 
         metadata = {"request_id": request_id, "reply_topic": self.reply_topic}
 
@@ -266,13 +346,39 @@ class BrokerRequestResponse(BrokerBase):
             "__solace_ai_connector_broker_request_reply_topic__"
         ] = self.reply_topic
 
-        encoded_payload = self.encode_payload(data["payload"])
+        if self.test_mode:
+            if self.broker_type == "test_streaming":
+                # The payload should be an array. Send one message per item in the array
+                if not isinstance(data["payload"], list):
+                    raise ValueError("Payload must be a list for test_streaming broker")
+                for item in data["payload"]:
+                    encoded_payload = self.encode_payload(item)
+                    self.pass_through_queue.put(
+                        Message(
+                            payload=encoded_payload,
+                            user_properties=deepcopy(data["user_properties"]),
+                            topic=data["topic"],
+                        )
+                    )
+            else:
+                encoded_payload = self.encode_payload(data["payload"])
+                self.pass_through_queue.put(
+                    Message(
+                        payload=encoded_payload,
+                        user_properties=data["user_properties"],
+                        topic=data["topic"],
+                    )
+                )
+        else:
+            encoded_payload = self.encode_payload(data["payload"])
+            self.messaging_service.send_message(
+                destination_name=data["topic"],
+                payload=encoded_payload,
+                user_properties=data["user_properties"],
+            )
 
-        self.messaging_service.send_message(
-            destination_name=data["topic"],
-            payload=encoded_payload,
-            user_properties=data["user_properties"],
-        )
+        data["stream"] = stream
+        data["streaming_complete_expression"] = streaming_complete_expression
 
         self.cache_service.add_data(
             key=request_id,
